backend/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==11.1.0
